logging added

+++ webApp/webApp/main.py
@@ -30,17 +30,17 @@ def dashboard():
 def add_post():
     form = PostForm()
     if form.validate_on_submit():
-        post = Post(title=form.title.data, content=form.content.data, author=form.author.data, slug=form.slug.data)
+        author = current_user.id
+        post = Post(title=form.title.data, content=form.content.data, author_id=author, slug=form.slug.data)
         # Clear form data
         form.title.data = ""
         form.content.data = ""
-        form.author.data = ""
         form.slug.data = ""
 
         #Add Post Data to Database
         db.session.add(post)
         db.session.commit()
-        current_app.logger.info(f"Author: {post.author} Post Added!")
+        current_app.logger.info(f"Author: {post.title} Post Added!")
         flash("Post Added!", category="info")
     
     return render_template("add-post.html", form=form)
@@ -58,7 +58,6 @@ def edit_post(id):
     form = PostForm()
     if form.validate_on_submit():
         post.title = form.title.data
-        post.author = form.author.data
         post.slug = form.slug.data
         post.content = form.content.data
         # Update Database
@@ -68,7 +67,6 @@ def edit_post(id):
         current_app.logger.info(f"{post.author} Post Edited Successfully")
         return redirect(url_for('main.post', id=post.id))
     form.title.data = post.title
-    form.author.data = post.author
     form.slug.data = post.slug
     form.content.data = post.content
     return render_template("edit-post.html", form=form)
@@ -77,17 +75,22 @@ def edit_post(id):
 @login_required
 def delete_post(id: int):
     post_to_delete = Post.query.get_or_404(id)
-    try:
-        db.session.delete(post_to_delete)
-        db.session.commit()
-        flash("Post Deleted Successfully", category="warning")
-        current_app.logger.info("Post Deleted Successfully")
-        posts = Post.query.order_by(Post.date_posted)
-        return render_template("main.html", posts=posts)
-    except:
-        flash("Post not Deleted", category="danger")
-        posts = Post.query.order_by(Post.date_posted)
-        return render_template("main.html", posts=posts)
+
+    if current_user.id == post_to_delete.author.id:
+        try:
+            db.session.delete(post_to_delete)
+            db.session.commit()
+            flash("Post Deleted Successfully", category="warning")
+            current_app.logger.info("Post Deleted Successfully")
+            posts = Post.query.order_by(Post.date_posted)
+            return render_template("main.html", posts=posts)
+        except:
+            flash("Post not Deleted", category="danger")
+            posts = Post.query.order_by(Post.date_posted)
+            return render_template("main.html", posts=posts)
+    else:
+        flash("You can't delete this post!", category="danger")
+
 
 @main.app_errorhandler(404)
 @login_required
